// src/app/model/representanteLegal.ts
export interface RepresentanteLegal {
    id: number;
    tipo_identificacion:   string;
    numero_identificacion: number;
    primer_apellido:       string;
    segundo_apellido:      string;
    primer_nombre:         string;
    segundo_nombre:        string;
    codigo_municipio:      string;
    telefono:              string;
    direccion:             string;
    email:                 string;
}